fix(solution): fix migrate v1 project programming language detection
